small fixes

+++ src/exceptions/price/PriceCalculatorTest.java
@@ -11,27 +11,27 @@ public class PriceCalculatorTest {
     public void ifFileIsReadableThenPriceIsCalculated() {
         MyFileResource.setData(2); // simulates that file contains number 2
 
-        Double price = new PriceCalculator().calculatePrice();
+        double price = new PriceCalculator().calculatePrice();
 
-        assertThat(price.intValue(), is(144));
+        assertThat(price, is(144));
     }
 
     @Test
     public void ifFileIsNotReadableThenPriceIsNotCalculated() {
         MyFileResource.setData(-1); // simulates read error
 
-        Double price = new PriceCalculator().calculatePrice();
+        double price = new PriceCalculator().calculatePrice();
 
-        assertThat(price.intValue(), is(-1));
+        assertThat(price, is(-1));
     }
 
     @Test
     public void ifFileIsNotReadableThenThrowException() {
         MyFileResource.makeItThrowOnRead(); // simulates exception on reading
 
-        Double price = new PriceCalculator().calculatePrice();
+        double price = new PriceCalculator().calculatePrice();
 
-        assertThat(price.intValue(), is(-1));
+        assertThat(price, is(-1));
     }
 
 }
